Restructure header and layout
