+++ src/App.js
@@ -1,22 +1,21 @@
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Layout from './components/Layout';
-import Rocket from './components/Rocket';
-import Mission from './components/Mission';
-import Profile from './components/Profile';
-import NoPage from './components/NoPage';
+import {
+  Rocket, Mission, Profile, NoPage,
+} from './components';
+import Layout from './components/shared/Layout';
 
 export default function App() {
   return (
     <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Layout />}>
-          <Route index element={<Rocket />} />
-          <Route path="mission" element={<Mission />} />
-          <Route path="profile" element={<Profile />} />
+      <Layout>
+        <Routes>
+          <Route index exact path="/" element={<Rocket />} />
+          <Route path="/missions" element={<Mission />} />
+          <Route path="/myprofile" element={<Profile />} />
           <Route path="*" element={<NoPage />} />
-        </Route>
-      </Routes>
+        </Routes>
+      </Layout>
     </BrowserRouter>
   );
 }
