+++ src/components/NoPage.js
@@ -1,3 +1,3 @@
-const NoPage = () => <h1>404</h1>;
+const NoPage = () => <h1>40</h1>;
 
 export default NoPage;
